+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+* Add `pooledByteBufferCounts` to ref data config.
+
 * Uplift LMDB to v0.8.1.
 
 * Changed implementaion of the byte buffer pool used in the ref data store to improve performance.